+++ GcodeCNCDemo4AxisRAMPS/GcodeCNCDemo4AxisRAMPS.ino
@@ -222,7 +222,7 @@ static float atan3(float dy,float dx) {
  **/
 float parseNumber(char code,float val) {
   char *ptr=serialBuffer;  // start at the beginning of buffer
-  while(ptr>1 && *ptr && ptr<serialBuffer+sofar) {  // walk to the end
+  while((long)ptr > 1 && (*ptr) && (long)ptr < (long)serialBuffer+sofar) {  // walk to the end
     if(*ptr==code) {  // if you find code on your walk,
       return atof(ptr+1);  // convert the digits that follow into a float and return it
     }
@@ -230,6 +230,7 @@ float parseNumber(char code,float val) {
   }
   return val;  // end reached, nothing found, return default val.
 }
+
 
 
 /**
